struct をpackageごとに定義していたので、modelで統一
